Split the signal_fir test suite by contract
R2.1 of the transform cleanup plan: signal_fir/tests.rs (4,324 lines,
74 tests) becomes signal_fir/tests/ grouped by contract (contract,
ui_tables, delays, recursion, placement, reverse_ad, coverage) with
shared pub(super) fixtures. Test names unchanged, transform suite stays
at 385, no production visibility widened.
