+++ crates/transform/src/signal_fir/vector/analysis/mod.rs
