+++ crates/transform/src/signal_fir/vector/assemble/mod.rs
