+++ crates/transform/src/signal_fir/vector/clock_ad/mod.rs
