+++ crates/transform/src/signal_fir/vector/module/mod.rs
